requirements-test.txt
-r requirements.txt

websocket-client    == 0.32.0
mock                == 1.3.0
coverage            == 4.0
nose                == 1.3.7
pylama              == 7.0.6